lock down deleteitem permissions

+++ src/resolvers/Mutation.js
@@ -37,10 +37,14 @@ const Mutations = {
 
   },
   
-  async deleteItem(parent, args, ctx, info) {
+  async deleteItem(parent, args, ctx, info) {   
     const where = { id: args.id };
-    const item = await ctx.db.query.item({ where }, `{ id title }`);
-    // TODO - verify user owns this item for deletion
+    const item = await ctx.db.query.item({ where }, `{ id title user { id } }`);
+    const ownsItem = item.user.id === ctx.request.userId;
+    const hasPermissions = ctx.request.user.permissions.some(permission => ['ADMIN', 'ITEMDELETE'].includes(permission));
+    if(!ownsItem && !hasPermissions) {
+      throw new Error ("You don't have permission to delete this item");
+    }
     return ctx.db.mutation.deleteItem({ where }, info);
   },
